Refactor verify tests
