+++ test/Sknet.OpenIddict.LiteDB.Tests/Resolvers/OpenIddictLiteDBTokenStoreResolverTests.cs
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace Sknet.OpenIddict.LiteDB.Tests;
+namespace Sknet.OpenIddict.LiteDB.Tests.Resolvers;
 
 public class OpenIddictLiteDBTokenStoreResolverTests
 {
