+++ test/Sknet.OpenIddict.LiteDB.Tests/Stores/OpenIddictLiteDBApplicationStoreTests.cs
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace Sknet.OpenIddict.LiteDB.Tests;
+namespace Sknet.OpenIddict.LiteDB.Tests.Stores;
 
 [UsesVerify]
 public class OpenIddictLiteDBApplicationStoreTests
