+++ test/Sknet.OpenIddict.LiteDB.Tests/Usings.cs
@@ -7,6 +7,7 @@ global using NCrunch.Framework;
 global using OpenIddict.Abstractions;
 global using OpenIddict.Core;
 global using Sknet.OpenIddict.LiteDB.Models;
+global using Sknet.OpenIddict.LiteDB.Tests.Builders;
 global using System.Collections.Generic;
 global using System.Collections.Immutable;
 global using System.Globalization;
@@ -18,3 +19,4 @@ global using Xunit.Abstractions;
 global using Xunit.Sdk;
 global using static Sknet.OpenIddict.LiteDB.Tests.Verifier;
 global using static VerifyXunit.Verifier;
+global using System.Runtime.CompilerServices;
