+++ test/Sknet.OpenIddict.LiteDB.Tests/Verify/Converters/JsonDocumentConverter.cs
@@ -1,0 +1,10 @@
+﻿/*
+ * MIT License
+ *
+ * Copyright (c) Simon Cropp
+ */
+class JsonDocumentConverter : WriteOnlyJsonConverter<JsonDocument>
+{
+    public override void Write(VerifyJsonWriter writer, JsonDocument value) =>
+        writer.Serialize(value.RootElement);
+}
