+++ test/Sknet.OpenIddict.LiteDB.Tests/Verify/Converters/JsonPropertyConverter.cs
@@ -1,0 +1,13 @@
+﻿/*
+ * MIT License
+ *
+ * Copyright (c) Simon Cropp
+ */
+class JsonPropertyConverter : WriteOnlyJsonConverter<JsonProperty>
+{
+    public override void Write(VerifyJsonWriter writer, JsonProperty value)
+    {
+        writer.WritePropertyName(value.Name);
+        writer.Serialize(value.Value);
+    }
+}
