+++ test/Sknet.OpenIddict.LiteDB.Tests/Verify/Verifier.cs
@@ -13,8 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-using System.Runtime.CompilerServices;
-
 namespace Sknet.OpenIddict.LiteDB.Tests;
 
 public static class Verifier
