+++ test/Sknet.OpenIddict.LiteDB.Tests/Verify/VerifySystemJson.cs
@@ -1,0 +1,19 @@
+﻿/*
+ * MIT License
+ *
+ * Copyright (c) Simon Cropp
+ */
+namespace Sknet.OpenIddict.LiteDB.Tests;
+
+public static class VerifySystemJson
+{
+    public static void Enable() =>
+        VerifierSettings
+            .AddExtraSettings(_ =>
+            {
+                var converters = _.Converters;
+                converters.Add(new JsonElementConverter());
+                converters.Add(new JsonDocumentConverter());
+                converters.Add(new JsonPropertyConverter());
+            });
+}
